+++ src/ui/ui_element.c
@@ -17,16 +17,14 @@
 /*!
  *  @file ui_element.c
  *
- *  @brief everything about drawing ui elements.
+ *  @brief initializing, setting up, drawing and attaching/detaching UI elements
+ *  to/from one another.
  */
 
-#include "../common/diagnostics.h"
-#include "../logger/logger.h"
-#include "../logger/logger_messages_internal.h"
-
-#include "ui.h"
-#include "ui_element_internal.h"
+#include "ui_element.h"
 #include "ui_types.h"
+
+#include <stddef.h>
 
 void fsl_ui_element_set_texture(fsl_ui_element *element, fsl_texture *texture)
 {
@@ -39,46 +37,63 @@ void fsl_ui_element_set_uv(fsl_ui_element *element,
         i32 uv_pos_x, i32 uv_pos_y, i32 uv_size_x, i32 uv_size_y)
 {
     element->flag |= FSL_FLAG_UI_DIRTY_TRANSFORM;
-    element->sprite.uv_pos.x = uv_pos_x;
-    element->sprite.uv_pos.y = uv_pos_y;
-    element->sprite.uv_size.x = uv_size_x;
-    element->sprite.uv_size.y = uv_size_y;
+    element->transform.uv_pos.x = uv_pos_x;
+    element->transform.uv_pos.y = uv_pos_y;
+    element->transform.uv_size.x = uv_size_x;
+    element->transform.uv_size.y = uv_size_y;
 }
 
 void fsl_ui_element_set_position(fsl_ui_element *element, i32 pos_x, i32 pos_y,
         i32 offset_x, i32 offset_y, i32 offset_scaled_x, i32 offset_scaled_y)
 {
     element->flag |= FSL_FLAG_UI_DIRTY_TRANSFORM;
-    element->sprite.pos.x = pos_x;
-    element->sprite.pos.y = pos_y;
-    element->sprite.offset.x = offset_x;
-    element->sprite.offset.y = offset_y;
-    element->sprite.offset_scaled.x = offset_scaled_x;
-    element->sprite.offset_scaled.y = offset_scaled_y;
+    element->transform.pos.x = pos_x;
+    element->transform.pos.y = pos_y;
+    element->transform.offset.x = offset_x;
+    element->transform.offset.y = offset_y;
+    element->transform.offset_scaled.x = offset_scaled_x;
+    element->transform.offset_scaled.y = offset_scaled_y;
 }
 
 void fsl_ui_element_set_size(fsl_ui_element *element,
         i32 size_x, i32 size_y, i32 size_scaled_x, i32 size_scaled_y)
 {
     element->flag |= FSL_FLAG_UI_DIRTY_TRANSFORM;
-    element->sprite.size.x = size_x;
-    element->sprite.size.y = size_y;
-    element->sprite.size_scaled.x = size_scaled_x;
-    element->sprite.size_scaled.y = size_scaled_y;
+    element->transform.size.x = size_x;
+    element->transform.size.y = size_y;
+    element->transform.size_scaled.x = size_scaled_x;
+    element->transform.size_scaled.y = size_scaled_y;
 }
 
 void fsl_ui_element_set_scale(fsl_ui_element *element, f32 scale_x, f32 scale_y)
 {
     element->flag |= FSL_FLAG_UI_DIRTY_TRANSFORM;
-    element->sprite.scale.x = scale_x;
-    element->sprite.scale.y = scale_y;
+    element->transform.scale.x = scale_x;
+    element->transform.scale.y = scale_y;
 }
 
 void fsl_ui_element_set_alignment(fsl_ui_element *element, i32 align_x, i32 align_y)
 {
     element->flag |= FSL_FLAG_UI_DIRTY_TRANSFORM;
-    element->sprite.align.x = align_x;
-    element->sprite.align.y = align_y;
+    element->transform.align.x = align_x;
+    element->transform.align.y = align_y;
+}
+
+void fsl_ui_element_set_9_slice(fsl_ui_element *element, b8 is_9_slice, i32 slice_size)
+{
+    if (is_9_slice)
+        element->flag |= FSL_FLAG_UI_9_SLICE;
+    else
+        element->flag &= ~FSL_FLAG_UI_9_SLICE;
+    element->transform.slice_size = slice_size;
+}
+
+void fsl_ui_element_set_callback(fsl_ui_element *element,
+        fsl_ui_event_type event_type,
+        void (*func)(fsl_ui_event event, void *data), void *data)
+{
+    element->callback[event_type].func = func;
+    element->callback[event_type].data = data;
 }
 
 void fsl_ui_element_attach(fsl_ui_element *parent, fsl_ui_element *child)
@@ -89,37 +104,12 @@ void fsl_ui_element_attach(fsl_ui_element *parent, fsl_ui_element *child)
     child->parent = parent;
 }
 
-void ui_element_bake_internal(fsl_ui_element *element)
+void fsl_ui_element_detach(fsl_ui_element *child)
 {
-    fsl_ui_sprite *s = &element->sprite;
-    v2f32 texture_scale = {0};
-    v2f32 texture_ratio = {0};
-    v2f32 size = {0};
-    v2f32 size_half = {0};
-    v2f32 alignment = {0};
-    v2f32 pos = {0};
+    if (child->parent)
+        child->parent->flag |= FSL_FLAG_UI_DIRTY_CHILDREN;
 
-    element->flag &= ~FSL_FLAG_UI_DIRTY_TRANSFORM;
-
-    texture_scale.x = 1.0f / element->texture->size.x;
-    texture_scale.y = 1.0f / element->texture->size.y;
-    texture_ratio.x = (f32)s->uv_size.x / element->texture->size.x;
-    texture_ratio.y = (f32)s->uv_size.y / element->texture->size.y;
-    size.x = (s->size.x + s->size_scaled.x * s->scale.x) * texture_ratio.x;
-    size.y = (s->size.y + s->size_scaled.y * s->scale.y) * texture_ratio.y;
-    size_half.x = size.x / 2.0f;
-    size_half.y = size.y / 2.0f;
-    alignment.x = size_half.x + s->align.x * size_half.x;
-    alignment.y = size_half.y + s->align.y * size_half.y;
-    pos.x = s->pos.x + s->offset.x + s->offset_scaled.x * s->scale.x;
-    pos.y = s->pos.y + s->offset.y + s->offset_scaled.y * s->scale.y;
-
-    s->uv_pos_baked.x = s->uv_pos.x * texture_scale.x;
-    s->uv_pos_baked.y = s->uv_pos.y * texture_scale.y;
-    s->uv_size_baked.x = s->uv_size.x * texture_scale.x;
-    s->uv_size_baked.y = s->uv_size.y * texture_scale.y;
-    s->pos_baked.x = pos.x - alignment.x;
-    s->pos_baked.y = pos.y - alignment.y;
-    s->size_baked.x = size.x;
-    s->size_baked.y = size.y;
+    child->flag |= FSL_FLAG_UI_DIRTY_PARENT;
+    child->flag |= FSL_FLAG_UI_DIRTY_TRANSFORM;
+    child->parent = NULL;
 }