+++ test.sh
@@ -1,2 +1,0 @@
-tar -cf fossil_engine-v0.10.0-beta-linux-x86_64.tar.gz fossil/
-zip -ry fossil_engine-v0.10.0-beta-linux-x86_64.zip fossil/