+++ tests/composable_ui/src/main.c
@@ -1,5 +1,6 @@
 #include "../../../fossil/deps/fossil/fossil_engine.h"
 
+fsl_texture *texture_p = NULL;
 fsl_render *render = {0};
 fsl_key_bind bind_quit = {0};
 fsl_key_bind bind_bigger_x = {0};
@@ -10,32 +11,94 @@ fsl_key_bind bind_align_left = {0};
 fsl_key_bind bind_align_right = {0};
 fsl_key_bind bind_align_up = {0};
 fsl_key_bind bind_align_down = {0};
+fsl_key_bind bind_element_attach = {0};
+fsl_key_bind bind_switch_workspace = {0};
 fsl_texture texture = {0};
-fsl_ui_element element = {0};
+fsl_ui_element element_parent = {0};
+fsl_ui_element element_child = {0};
+fsl_ui_element element_panel = {0};
 v2f32 scale = {0};
 v2i32 align = {0};
 
 void ui_update(void)
 {
+    fsl_ui_element_set_position(&element_parent, render->mouse_pos.x, render->mouse_pos.y, 0, 0, 0, 0);
+    fsl_ui_element_set_scale(&element_parent, scale.x, scale.y);
+    fsl_ui_element_set_alignment(&element_parent, align.x, align.y);
+}
+
+void ui_draw_1(void)
+{
     fsl_ui_start(FALSE, TRUE);
-    fsl_ui_element_set_position(&element, render->mouse_pos.x, render->mouse_pos.y, 0, 0, 0, 0);
-    fsl_ui_element_set_scale(&element, scale.x, scale.y);
-    fsl_ui_element_set_alignment(&element, align.x, align.y);
-
-    fsl_ui_element_draw(&element);
+    fsl_ui_element_draw(&element_parent);
+    fsl_ui_element_draw(&element_child);
     fsl_ui_stop();
-
     fsl_fbo_blit(0);
+}
+
+void ui_draw_2(void)
+{
+    fsl_ui_start(FALSE, TRUE);
+    fsl_ui_element_draw(&element_panel);
+    fsl_ui_stop();
+    fsl_fbo_blit(0);
+}
+
+void resize(fsl_ui_event event, void *data)
+{
+    (void)data;
+    fsl_ui_element_set_scale(event.caller, scale.x, scale.y);
+    LOGSUCCESS(0, "ENTER\n");
+}
+
+void move(fsl_ui_event event, void *data)
+{
+    static i32 i = 0;
+
+    (void)data;
+    fsl_ui_element_set_position(event.caller, i, 0, 0, 0, 0, 0);
+    ++i;
+
+    LOGSUCCESS(0, fsl_logger_stringf("HOVER: %d\n", i));
+}
+
+void reset(fsl_ui_event event, void *data)
+{
+    (void)data;
+    fsl_ui_element_set_position(event.caller, 0, 0, 0, 0, 0, 0);
+
+    LOGSUCCESS(0, "LEAVE: 0\n");
+}
+
+void fuck(fsl_ui_event event, void *data)
+{
+    v2f64 pos = *(v2f64*)data;
+    fsl_ui_element_set_position(event.caller, pos.x, pos.y, 0, 0, 0, 0);
+
+    LOGSUCCESS(0, "CLICK\n");
+}
+
+void enter(fsl_ui_event event, void *data)
+{
+    fsl_ui_element_set_texture(event.caller, &texture_p[FSL_TEXTURE_INDEX_PANEL_ACTIVE]);
+}
+
+void leave(fsl_ui_event event, void *data)
+{
+    fsl_ui_element_set_texture(event.caller, &texture_p[FSL_TEXTURE_INDEX_PANEL_INACTIVE]);
 }
 
 int main(int argc, char **argv)
 {
+
     if (fsl_engine_init(argc, argv, NULL, 1280, 720, 0) != FSL_ERR_SUCCESS)
         goto cleanup;
 
     if (fsl_texture_init(&texture, "Panel", "panel", "panel.png", "assets/",
                 GL_RGBA, GL_NEAREST, FSL_COLOR_CHANNELS_RGBA, FALSE, FALSE) != FSL_ERR_SUCCESS)
         goto cleanup;
+
+    texture_p = fsl_mem_handle_get(fsl_texture_buf);
 
     render = fsl_render_get();
 
@@ -48,45 +111,141 @@ int main(int argc, char **argv)
     bind_align_right = fsl_key_bind_init(FSL_KEY_RIGHT, 0, 0, 0, 0, 0);
     bind_align_up = fsl_key_bind_init(FSL_KEY_UP, 0, 0, 0, 0, 0);
     bind_align_down = fsl_key_bind_init(FSL_KEY_DOWN, 0, 0, 0, 0, 0);
+    bind_element_attach = fsl_key_bind_init(FSL_KEY_ENTER, 0, 0, 0, 0, 0);
+    bind_switch_workspace = fsl_key_bind_init(FSL_KEY_SPACE, 0, 0, 0, 0, 0);
 
     scale.x = 1.0f;
     scale.y = 1.0f;
     align.x = -1;
     align.y = -1;
 
-    fsl_ui_element_set_texture(&element, &texture);
-    fsl_ui_element_set_uv(&element, 0, 0, 177, 177);
-    fsl_ui_element_set_position(&element, 0, 0, 0, 0, 0, 0);
-    fsl_ui_element_set_size(&element, 0, 0, texture.size.x, texture.size.x);
-    fsl_ui_element_set_scale(&element, scale.x, scale.y);
-    fsl_ui_element_set_alignment(&element, align.x, align.y);
+    fsl_ui_element_set_texture(&element_parent, &texture);
+    fsl_ui_element_set_uv(&element_parent, 0, 0, 177, 177);
+    fsl_ui_element_set_position(&element_parent, 0, 0, 0, 0, 0, 0);
+    fsl_ui_element_set_size(&element_parent, 0, 0, 177, 177);
+    fsl_ui_element_set_scale(&element_parent, scale.x, scale.y);
+    fsl_ui_element_set_alignment(&element_parent, align.x, align.y);
+
+    fsl_ui_element_set_texture(&element_child, &texture);
+    fsl_ui_element_set_uv(&element_child, 0, 0, 177, 177);
+    fsl_ui_element_set_position(&element_child, 0, 0, 0, 0, 0, 0);
+    fsl_ui_element_set_size(&element_child, 0, 0, texture.size.x, texture.size.y);
+    fsl_ui_element_set_scale(&element_child, 1.0f, 1.0f);
+    fsl_ui_element_set_alignment(&element_child, -1, -1);
+    fsl_ui_element_set_callback(&element_child, FSL_UI_EVENT_TYPE_ENTER, &resize, NULL);
+    fsl_ui_element_set_callback(&element_child, FSL_UI_EVENT_TYPE_HOVER, &move, NULL);
+    fsl_ui_element_set_callback(&element_child, FSL_UI_EVENT_TYPE_LEAVE, &reset, NULL);
+    fsl_ui_element_set_callback(&element_child, FSL_UI_EVENT_TYPE_CLICK, &fuck, &render->mouse_pos);
+
+    fsl_ui_element_set_uv(&element_panel, 0, 0, 16, 16);
+    fsl_ui_element_set_position(&element_panel, 10, 10, 0, 0, 0, 0);
+    fsl_ui_element_set_size(&element_panel, 0, render->size.y - 20, 400, 0);
+    fsl_ui_element_set_scale(&element_panel, 1.0f, 1.0f);
+    fsl_ui_element_set_alignment(&element_panel, -1, -1);
+    fsl_ui_element_set_9_slice(&element_panel, TRUE, 8);
+    fsl_ui_element_set_texture(&element_panel, &texture_p[FSL_TEXTURE_INDEX_PANEL_INACTIVE]);
+    fsl_ui_element_set_callback(&element_panel, FSL_UI_EVENT_TYPE_ENTER, &enter, NULL);
+    fsl_ui_element_set_callback(&element_panel, FSL_UI_EVENT_TYPE_LEAVE, &leave, NULL);
+
+workspace_1:
 
     while (fsl_engine_running(NULL))
     {
         glClearColor(0.1f, 0.1f, 0.1f, 1.0f);
-
-        ui_update();
+        if (render->mouse_delta.x || render->mouse_delta.y)
+            ui_update();
 
         if (fsl_is_key_hold(bind_bigger_x))
+        {
             scale.x += 0.1f;
+            ui_update();
+        }
 
         if (fsl_is_key_hold(bind_smaller_x))
+        {
             scale.x -= 0.1f;
+            ui_update();
+        }
 
         if (fsl_is_key_hold(bind_bigger_y))
+        {
             scale.y += 0.1f;
+            ui_update();
+        }
 
         if (fsl_is_key_hold(bind_smaller_y))
+        {
             scale.y -= 0.1f;
+            ui_update();
+        }
 
         if (fsl_is_key_press(bind_align_left))
+        {
             align.x += 1;
+            ui_update();
+        }
+
         if (fsl_is_key_press(bind_align_right))
+        {
             align.x -= 1;
+            ui_update();
+        }
+
         if (fsl_is_key_press(bind_align_up))
+        {
             align.y += 1;
+            ui_update();
+        }
+
         if (fsl_is_key_press(bind_align_down))
+        {
             align.y -= 1;
+            ui_update();
+        }
+
+        if (fsl_is_key_press(bind_element_attach))
+        {
+            if (element_child.parent)
+                fsl_ui_element_detach(&element_child);
+            else
+                fsl_ui_element_attach(&element_parent, &element_child);
+
+            ui_update();
+        }
+
+        ui_draw_1();
+
+        if (fsl_is_key_press(bind_switch_workspace))
+            goto workspace_2;
+
+        if (fsl_is_key_press(bind_quit))
+            fsl_request_engine_close();
+
+        fsl_limit_framerate(60, render->time);
+    }
+
+workspace_2:
+
+    while (fsl_engine_running(NULL))
+    {
+        glClearColor(0.1f, 0.1f, 0.1f, 1.0f);
+        if (render->mouse_delta.x || render->mouse_delta.y)
+            ui_update();
+
+        if (fsl_is_key_press(bind_element_attach))
+        {
+            if (element_panel.parent)
+                fsl_ui_element_detach(&element_panel);
+            else
+                fsl_ui_element_attach(&element_parent, &element_panel);
+
+            ui_update();
+        }
+
+        ui_draw_2();
+
+        if (fsl_is_key_press(bind_switch_workspace))
+            goto workspace_1;
 
         if (fsl_is_key_press(bind_quit))
             fsl_request_engine_close();