+++ tests/game_hhc/src/main.c
@@ -148,7 +148,7 @@ static u32 settings_init(void)
 
     settings.lerp_speed = SET_LERP_SPEED_DEFAULT;
 
-    settings.render_distance = 16;
+    settings.render_distance = SET_RENDER_DISTANCE_DEFAULT;
     settings.chunk_buf_radius = settings.render_distance;
     settings.chunk_buf_diameter = settings.chunk_buf_radius * 2 + 1;
 
@@ -192,7 +192,7 @@ void settings_update(void)
     fsl_ui_element_set_texture(&element_item_bar, &texture_p[TEXTURE_ITEM_BAR]);
     fsl_ui_element_set_uv(&element_item_bar, 0, 0, 169, 16);
     fsl_ui_element_set_position(&element_item_bar, render->size.x / 2, render->size.y, 0, 0, 0, -2);
-    fsl_ui_element_set_size(&element_item_bar, 0, 0, element_item_bar.texture->size.x, element_item_bar.texture->size.y);
+    fsl_ui_element_set_size(&element_item_bar, 0, 0, 169, 16);
     fsl_ui_element_set_scale(&element_item_bar, settings.gui_scale, settings.gui_scale);
     fsl_ui_element_set_alignment(&element_item_bar, 0, 1);
 }
@@ -593,7 +593,7 @@ static void draw_everything(void)
 
      glUniformMatrix4fv(uniform.bounding_box.mat_perspective, 1, GL_FALSE,
              (GLfloat*)&player.camera.projection.perspective);
- 
+
     /* ---- draw player chunk bounding box ---------------------------------- */
 
     if (core.debug.chunk_bounds)