+++ CHANGELOG.md
@@ -1,11 +1,59 @@
 # changelog
 
 >**NOTE:**
->- history is mixed with the repo "awertnex/heaven-hell_continuum" since
-   this engine originated from the making of that game and later was moved here
-   at engine version v0.3.3-beta (2026 01 24), which makes v0.3.4-beta the
-   first version of the engine as standalone in this repo
+>- history is mixed with the repo
+   "https://github.com/awertnex/heaven-hell_continuum.git" since this engine
+   originated from the making of that game and later was moved here at engine
+   version v0.3.3-beta (2026 01 24), which makes v0.3.4-beta the first version
+   of the engine as standalone in this repo
 
+## v0.11.0-beta (2026 06 12)
+
+#### changes
+- changed size parameter in UI element drawing from UV-relative to constant
+  pixel size.
+- changed parameter `usage` in function `fsl_mesh_generate()` to `draw_type`,
+  with type `fsl_draw_type`.
+- changed UI drawing from using uniforms to pass position and size to uploading
+  position and size to a dynamic vertex buffer.
+
+#### additions
+- added UI element attachment and position following of parent.
+- added UI element detachment from parent.
+- added UI event listening and callback attachment to elements:
+    - added ENTER, HOVER, LEAVE, CLICK, HOLD and RELEASE event listening and
+      handling.
+    - added function `fsl_ui_element_set_callback()` for attaching callbacks to
+      elements (listening happens within function `fsl_ui_element_draw()`).
+- added UI 9-slice setter function `fsl_ui_element_set_9_slice()`.
+- added support for drawing 9-slice using the same functions and struct as a
+  regular element.
+- exposed function `fsl_ui_element_bake()` to manually bake/refresh element
+  parameters (also called automatically from function `fsl_ui_element_draw()` if
+  transform dirty).
+
+#### deletions
+- removed deprecations from v0.9.0-beta and earlier:
+    - struct `fsl_mesh` parameters `vbo`, `vbo_len`, `vbo_data`, `ebo`,
+      `ebo_len` and `ebo_data`.
+    - function `fsl_get_camera_lookat_angles()`.
+    - shader 'ui_deprecated'.
+
+#### deprecations
+- deprecated parameter `nine_slice` in function `fsl_ui_start()`.
+- deprecated shader `ui_9_slice`, shader `ui` will be used instead.
+- deprecated functions `fsl_key_bind_attach()` and `fsl_input_context_set()`.
+- deprecated parameter `context` in function `fsl_key_bind_init()`.
+
+#### bugs and flaws
+- the 'obj' loader possibly has a bug due to hash-table collisions:
+    - sometimes faces disappear when ordered in specific ways in the obj file.
+- the 'obj' loader does not handle concave faces smartly, just loads them as
+  they are.
+- game UI is clunky right now because new UI API and didn't bother to change it,
+  yet.
+
+- - -
 ## v0.10.1-beta (2026 06 10)
 
 #### additions